requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf
pymupdf
sentence-transformers
chromadb
langchain-groq
python-dotenv
fastapi
uvicorn
